Widen sample_counts_ uint16->uint32 (remove 65535 spp ceiling)
Per-pixel Welford sample count was uint16_t, wrapping at 65536 spp and
flooring RMSE for very-high-spp reference renders. Widen the device POD
pointer, the host AsyncBuffer, and the raygen write cast to uint32_t
(~4.29e9 ceiling). Bit-identical for spp<=65535 (cast yields the same
value); furnace PASS. Cost: +2 bytes/pixel (~1MB at 900x600), negligible.

Key changes:
- device/params/image.h: sample_counts_ uint16_t* -> uint32_t*
- host/params/image.h: AsyncBuffer<uint16_t> -> <uint32_t>, const_cast update
- raygen.cuh: write-back cast uint16_t -> uint32_t

+++ device/entry/raygen.cuh
@@ -274,7 +274,7 @@ extern "C" __global__ void __raygen__rg() {
         launch_params.image_.variance_[pixel_linear_idx] = make_float4(M2);
     }
     launch_params.image_.mean_[pixel_linear_idx] = make_float4(mean);
-    launch_params.image_.sample_counts_[pixel_linear_idx] = static_cast<uint16_t>(
+    launch_params.image_.sample_counts_[pixel_linear_idx] = static_cast<uint32_t>(
         prev_count + launch_params.image_.batch_size_);
     if (launch_params.image_.albedo_aov_) {
         launch_params.image_.albedo_aov_[pixel_linear_idx] = make_float4(aov_albedo);
